Removing the Lexer module so stuff gets inlined
