+++ package.yaml
@@ -53,7 +53,6 @@ library:
   - Language.Docker.PrettyPrint
   - Language.Docker.Normalize
   - Language.Docker.Syntax
-  - Language.Docker.Lexer
   - Language.Docker.Syntax.Lift
   - Language.Docker.EDSL
   - Language.Docker.EDSL.Quasi
