+++ test/Language/Docker/ParserSpec.hs
@@ -255,29 +255,29 @@ spec = do
                             ]
                 in normalizeEscapedLines longEscapedCmd `shouldBe` longEscapedCmdExpected
         describe "expose" $ do
-            it "should handle number ports" $ do
+            it "should handle number ports" $
                 let content = "EXPOSE 8080"
-                parse expose "" content `shouldBe` Right (Expose (Ports [Port 8080 TCP]))
-            it "should handle many number ports" $ do
+                in assertAst content [Expose (Ports [Port 8080 TCP])]
+            it "should handle many number ports" $
                 let content = "EXPOSE 8080 8081"
-                parse expose "" content `shouldBe` Right (Expose (Ports [Port 8080 TCP, Port 8081 TCP]))
-            it "should handle ports with protocol" $ do
+                in  assertAst content [Expose (Ports [Port 8080 TCP, Port 8081 TCP])]
+            it "should handle ports with protocol" $
                 let content = "EXPOSE 8080/TCP 8081/UDP"
-                parse expose "" content `shouldBe` Right (Expose (Ports [Port 8080 TCP, Port 8081 UDP]))
-            it "should handle ports with protocol and variables" $ do
+                in assertAst content [Expose (Ports [Port 8080 TCP, Port 8081 UDP])]
+            it "should handle ports with protocol and variables" $
                 let content = "EXPOSE $PORT 8080 8081/UDP"
-                parse expose "" content `shouldBe` Right (Expose (Ports [PortStr "$PORT", Port 8080 TCP, Port 8081 UDP]))
-            it "should handle port ranges" $ do
+                in assertAst content [Expose (Ports [PortStr "$PORT", Port 8080 TCP, Port 8081 UDP])]
+            it "should handle port ranges" $
                 let content = "EXPOSE 80 81 8080-8085"
-                parse expose "" content `shouldBe` Right (Expose (Ports [Port 80 TCP, Port 81 TCP, PortRange 8080 8085 TCP]))
-            it "should handle udp port ranges" $ do
+                in assertAst content [Expose (Ports [Port 80 TCP, Port 81 TCP, PortRange 8080 8085 TCP])]
+            it "should handle udp port ranges" $
                 let content = "EXPOSE 80 81 8080-8085/udp"
-                parse expose "" content `shouldBe` Right (Expose (Ports [Port 80 TCP, Port 81 TCP, PortRange 8080 8085 UDP]))
+                in assertAst content [Expose (Ports [Port 80 TCP, Port 81 TCP, PortRange 8080 8085 UDP])]
 
         describe "syntax" $ do
-            it "should handle lowercase instructions (#7 - https://github.com/beijaflor-io/haskell-language-dockerfile/issues/7)" $ do
+            it "should handle lowercase instructions (#7 - https://github.com/beijaflor-io/haskell-language-dockerfile/issues/7)" $
                 let content = "from ubuntu"
-                parse dockerfile "" content `shouldBe` Right [InstructionPos (From (UntaggedImage "ubuntu" Nothing)) "" 1]
+                in assertAst content [From (UntaggedImage "ubuntu" Nothing)]
 
         describe "ADD" $ do
             it "simple ADD" $
